spark/requirements.txt
pyspark==3.5.0
pandas==2.2.2
pyarrow==16.1.0
entsoe-py==0.7.11
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
python-dotenv==1.0.1
tenacity==8.3.0
